a unified utils.py
