+++ akd_ext/agents/_base/pydantic_ai/_base.py
@@ -51,9 +51,7 @@ from akd._base.protocols import AKDExecutable, RunContextProtocol
 from akd._base.structures import RunContext as AKDRunContext
 from akd.agents._base import BaseAgentConfig
 from akd.tools._base import BaseTool
-
-from ._event_translator import pai_event_to_akd_event
-from ._tool_adapter import akd_to_pai_tool
+from ._utils import akd_to_pai_tool, pai_event_to_akd_event
 
 # ---------------------------------------------------------------------------
 # Config
